+++ index.js
@@ -1,7 +1,12 @@
+const store = require('./src/store');
 const Normalizer = require('./src/normalizer');
-const DataPackageService = require('./src/service');
 
-const dataPackageService = new DataPackageService();
+const dataPackageService = store.dataPackageService;
+
+dataPackageService.Normalizer = Normalizer;
+dataPackageService.Resource = store.Resource;
+dataPackageService.Package = store.Package;
+dataPackageService.makeResource = store.makeResource;
+dataPackageService.makePackage = store.makePackage;
 
 module.exports = dataPackageService;
-module.exports.Normalizer = Normalizer;